Module Heft

module Derived : sig ... end

module Elaborator : sig ... end

module Inductive : sig ... end

Machinery for defining inductive types without large amounts of manual derivation.

module Kernel : sig ... end

module Lexer : sig ... end

module Parser : sig ... end

AST types

module Printing : sig ... end

module Rewrite : sig ... end

module Rules : sig ... end

module Synth : sig ... end

Type-directed term enumeration.

module Tactic : sig ... end

Tactic engine for interactive theorem proving with algebraic effects.

module Theories : sig ... end

Example definitions used in some test proofs

module Util : sig ... end