Module Heft.Derived

Derived rules and some basic definitions for use in tactics

val term_subst : Kernel.term -> Kernel.term -> Kernel.term -> Kernel.term

Replace all occurrences of old_tm with new_tm in tm

val term_size : Kernel.term -> int

val make_exn : ('a, 'b) Stdlib.result -> 'a

val init_types : unit -> (unit, [> `TypeAlreadyDeclared of string ]) Stdlib.result

val init_true : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_true : unit -> Kernel.term

val init_forall : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_forall_const : Kernel.hol_type -> Kernel.term

val make_forall : Kernel.term -> Kernel.term -> Kernel.term

val make_foralls : Kernel.term list -> Kernel.term -> Kernel.term

val init_conj : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_conj_const : unit -> Kernel.term

val make_conj : Kernel.term -> Kernel.term -> Kernel.term

val make_conjs : Kernel.term list -> Kernel.term

val init_imp : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_imp_const : unit -> Kernel.term

val make_imp : Kernel.term -> Kernel.term -> Kernel.term

val make_imps : Kernel.term list -> Kernel.term -> Kernel.term

val init_false : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_false : unit -> Kernel.term

val init_neg : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_neg_const : unit -> Kernel.term

val make_neg : Kernel.term -> Kernel.term

val init_exists : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_exists_const : Kernel.hol_type -> Kernel.term

val make_exists : Kernel.term -> Kernel.term -> Kernel.term

val init_disj : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_disj_const : unit -> Kernel.term

val make_disj : Kernel.term -> Kernel.term -> Kernel.term

val make_neq : Kernel.term -> Kernel.term -> (Kernel.term, [> `CantApplyNonFunctionType of Kernel.term | `UnexpectedLambdaForm ]) Stdlib.result

val init_classical : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `NewAxiomNotAProp | `UnexpectedLambdaForm ]) Stdlib.result

val make_select_const : unit -> Kernel.term

val make_select : Kernel.term -> Kernel.term -> (Kernel.term, [> `MakeLamNotAVariable of Kernel.term ]) Stdlib.result

val init_choice : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `MakeAppTypesDontAgree of Kernel.hol_type * Kernel.hol_type | `MakeLamNotAVariable of Kernel.term | `NewAxiomNotAProp | `UnexpectedLambdaForm ]) Stdlib.result

val init_cond : unit -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val make_cond_const : Kernel.hol_type -> Kernel.term

val make_cond : Kernel.term -> Kernel.term -> Kernel.term -> Kernel.term

val reset : unit -> (unit, 'a) Stdlib.result

val true_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

T = ((λp:bool. p) = (λp. p))

val forall_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

∀ = (λP. P = (λx. T))

val conj_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

∧ = (λp q. (λf. f p q) = (λf. f T T))

val imp_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

==> = (λp q. p ∧ q ⟺ p)

val false_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

F = (∀p. p)

val neg_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

¬ = (λp. p ==> F)

val exists_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

∃ = (λP. ¬(∀x. ¬(P x)))

val disj_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

∨ = (λp q. ∀r. (p ==> r) ==> (q ==> r) ==> r)

val classical_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `NewAxiomNotAProp | `UnexpectedLambdaForm ]) Stdlib.result

∀p. p /\ ¬p

val choice_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `MakeAppTypesDontAgree of Kernel.hol_type * Kernel.hol_type | `MakeLamNotAVariable of Kernel.term | `NewAxiomNotAProp | `UnexpectedLambdaForm ]) Stdlib.result

val cond_def : (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `UnexpectedLambdaForm ]) Stdlib.result

val destruct_conj : Kernel.term -> (Kernel.term * Kernel.term, [> `NotAConj ]) Stdlib.result

val is_conj : Kernel.term -> bool

val is_eq : Kernel.term -> bool

val is_disj : Kernel.term -> bool

val destruct_imp : Kernel.term -> (Kernel.term * Kernel.term, [> `NotAnImp ]) Stdlib.result

val destruct_disj : Kernel.term -> (Kernel.term * Kernel.term, [> `NotADisj ]) Stdlib.result

val is_imp : Kernel.term -> bool

val destruct_exists : Kernel.term -> (Kernel.term * Kernel.term, [> `NotAnExists ]) Stdlib.result

val type_of_existential : ('a * Kernel.term) -> (Kernel.hol_type, [> `CantApplyNonFunctionType of Kernel.term | `NotAnExists | `UnexpectedLambdaForm ]) Stdlib.result

val term_of_negation : Kernel.term -> (Kernel.term, [> `NotANegation ]) Stdlib.result

val is_neg : Kernel.term -> bool

val destruct_forall : Kernel.term -> (Kernel.term * Kernel.term, [> `NotAForall ]) Stdlib.result

val quantifier_of_forall : Kernel.term -> (Kernel.term, [> `NotAForall ]) Stdlib.result

val body_of_forall : Kernel.term -> (Kernel.term, [> `NotAForall ]) Stdlib.result

type side =

| Left

| Right

val side_of_op : string -> side -> Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

val conj_left_term : Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

obtain p from p /\ q

val conj_right_term : Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

obtain q from p /\ q

val disj_left_term : Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

obtain p from p \/ q

val disj_right_term : Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

obtain q from p \/ q

val imp_left_term : Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

obtain p from p ==> q

val imp_right_term : Kernel.term -> (Kernel.term, [> `OperationDoesntMatch of string ]) Stdlib.result

obtain q from p ==> q

val ap_term : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `NotBothEquations | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val ap_thm : Kernel.thm -> Kernel.term -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `NotBothEquations | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val sym : Kernel.thm -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `CantDestructEquality | `Eq_MP of Kernel.term | `NotAnApplication of Kernel.term | `NotBothEquations | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- x = y should derive |- y = x

val truth : Kernel.thm

λp. p = λp. p is truth

val eq_truth_intro : Kernel.thm -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `UnexpectedLambdaForm ]) Stdlib.result

|- p should derive |- p = T

val eq_truth_elim : Kernel.thm -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `CantDestructEquality | `Eq_MP of Kernel.term | `NotAnApplication of Kernel.term | `NotBothEquations | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- p = T should derive |- p

val beta_conv : Kernel.term -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `NotALam | `NotAnApp | `NotTrivialBetaRedex | `UnexpectedLambdaForm ]) Stdlib.result

|- (λx. x /\ x) arg should derive |- arg /\ arg

val rhs : Kernel.thm -> (Kernel.term, [> `CantDestructEquality ]) Stdlib.result

val lhs : Kernel.thm -> (Kernel.term, [> `CantDestructEquality ]) Stdlib.result

val deep_beta : Kernel.term -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `LamRuleCantApply | `NotALam | `NotAnApp | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

recursively apply beta conversions

val conv_left : (Kernel.term -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `CantDestructEquality | `Eq_MP of Kernel.term | `NotAnApplication of Kernel.term | `NotBothEquations | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ] as 'a) Stdlib.result) -> Kernel.thm -> (Kernel.thm, 'a) Stdlib.result

obtain equality after applying a conversion to the left side

val conv_right : (Kernel.term -> (Kernel.thm, [> `CantDestructEquality | `RuleTrans ] as 'a) Stdlib.result) -> Kernel.thm -> (Kernel.thm, 'a) Stdlib.result

obtain equality after applying a conversion to the right side

val conv_equality : (Kernel.term -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `CantDestructEquality | `Eq_MP of Kernel.term | `NotAnApplication of Kernel.term | `NotBothEquations | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ] as 'a) Stdlib.result) -> Kernel.thm -> (Kernel.thm, 'a) Stdlib.result

obtain equality after applying a converion to both sides

val unfold_definition : Kernel.thm -> Kernel.term list -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Eq_MP of Kernel.term | `NotALam | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val conj : Kernel.thm -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- p, |- q should derive |- p /\ q

val conj_lr : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val conj_left : Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- p /\ q should derive |- p

val conj_right : Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val undisch : Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAProposition | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- p => q should derive {p} |- q

val disch : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAProposition | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

{p} |- q should derive |- p => q

val prove_hyp : Kernel.thm -> Kernel.thm -> (Kernel.thm, [> `CantApplyNonFunctionType of Kernel.term | `Eq_MP of Kernel.term | `UnexpectedLambdaForm ]) Stdlib.result

val mp : Kernel.thm -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAProposition | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- p => q, |- p should derive |- q

val gen : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- x = x should derive |- ∀x. x = x when x is not free in hypotheses

val gens : Kernel.term list -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val spec : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- ∀x. x = x should derive |- t = t for any term t

val specs : Kernel.term list -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

val disj_left : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ P should derive ⊢ P ∨ Q

val disj_right : Kernel.thm -> Kernel.term -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ Q should derive ⊢ P ∨ Q

val disj_cases : Kernel.thm -> Kernel.thm -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ P ∨ Q, {P} ⊢ R, {Q} ⊢ R should derive ⊢ R

val not_elim : Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotANegation | `NotAProposition | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ ¬P should derive {P} ⊢ F

val not_intro : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotAProposition | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

{P} ⊢ F should derive ⊢ ¬P

val contr : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ F should derive ⊢ P (ex falso quodlibet)

val ccontr : Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewAxiomNotAProp | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotANegation | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

{¬P} ⊢ F should derive ⊢ P (classical contradiction)

val exists_p : Kernel.term -> Kernel.term -> Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotANegation | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ t = t should derive ⊢ ∃x. x = x

val exists : Kernel.term -> Kernel.term -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotANegation | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ t = t should derive ⊢ ∃x. x = x

val choose : Kernel.term -> Kernel.thm -> Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `Clash of Kernel.term | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewAxiomNotAProp | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotAForall | `NotALam | `NotANegation | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

⊢ ∃x. P(x), {P(x)} ⊢ Q should derive ⊢ Q (where x not free in Q)

val neg_sym : Kernel.thm -> (Kernel.thm, [> `BadSubstitutionList | `CantApplyNonFunctionType of Kernel.term | `CantCreateVariantForNonVariable of Kernel.term | `CantDestructEquality | `ConstantTermAlreadyDeclared of string | `DefinitionError of string | `Eq_MP of Kernel.term | `LamRuleCantApply | `MakeLamNotAVariable of Kernel.term | `NewBasicDefinition of Kernel.term | `NewBasicDefinitionAlreadyDefined of string | `NotALam | `NotANegation | `NotAProposition | `NotAnApp | `NotAnApplication of Kernel.term | `NotBothEquations | `NotTrivialBetaRedex | `OperationDoesntMatch of string | `RuleTrans | `TypesDontAgree | `UnexpectedLambdaForm ]) Stdlib.result

|- ~ (x = y) should derive |- ~(y = x)

val bool_cases : Kernel.thm