Module Heft.Grimm

type Stdlib.Effect.t +=

| CurrentProb : float Stdlib.Effect.t

val prob_of_tactic : Tactic.tactic -> Tactic.goal -> string * float

type choiceKind =

| CTerm of Kernel.term

| CTheorem of Kernel.thm

| CTactic of Tactic.tactic

| CUnknown

an expansion represents the search space. An edge is a potential tactic*goal combo to expand, and a choice is any other choice

val pp_choiceKind : Ppx_deriving_runtime.Format.formatter -> choiceKind -> Ppx_deriving_runtime.unit

val show_choiceKind : choiceKind -> Ppx_deriving_runtime.string

type expansion =

| Edge of string

| Choice of choiceKind

| Done of Kernel.thm

val pp_expansion : Ppx_deriving_runtime.Format.formatter -> expansion -> Ppx_deriving_runtime.unit

val show_expansion : expansion -> Ppx_deriving_runtime.string

val thm_of_expansion : expansion -> Kernel.thm option

type node = {

root_tactic : Tactic.tactic;

down : (unit -> node list) option;

up : (Kernel.thm -> node list) option;

expansion : expansion;

goal : Tactic.goal;

id : Uuidm.t;

prob : float;

}

A node has an expansion type as well as some other data.

val pp_node : Ppx_deriving_runtime.Format.formatter -> node -> Ppx_deriving_runtime.unit

val show_node : node -> Ppx_deriving_runtime.string

val uuid : unit -> Uuidm.t

val test_id : Uuidm.t

module Priority : sig ... end

module Frontier : sig ... end

val make_node : ?down:(unit -> node list) -> ?up:(Kernel.thm -> node list) -> Tactic.tactic -> Tactic.goal -> Uuidm.t -> float -> expansion -> node

val with_prob : float -> (unit -> 'a) -> 'a

val expand : (Uuidm.t, unit) Stdlib.Hashtbl.t -> node -> node list

val search : (Uuidm.t, unit) Stdlib.Hashtbl.t -> Frontier.t -> int -> Kernel.thm

val frontier_of_goal : Tactic.tactic -> Tactic.goal -> Frontier.t

val with_grimm : ?depth:int -> Tactic.tactic_combinator

val gauto : Tactic.tactic