Module Heft.Tactic

type goal = Kernel.term list * Kernel.term

a list of assumptions and a term to prove under them

val pp_goal : Ppx_deriving_runtime.Format.formatter -> goal -> Ppx_deriving_runtime.unit

val show_goal : goal -> Ppx_deriving_runtime.string

type level =

| Debug

| Info

| Warn

| Error

| Proof

| Search

level is used to distinguish between different types of traces

type proof_state =

| Incomplete of goal

| Complete of Kernel.thm

proof_state is used by the ambient handler prove to represent the result of applying a tactic

val pp_proof_state : Ppx_deriving_runtime.Format.formatter -> proof_state -> Ppx_deriving_runtime.unit

val show_proof_state : proof_state -> Ppx_deriving_runtime.string

type tactic = goal -> Kernel.thm

a tactic is a function that works on a goal, possibly performing effects

type tactic_combinator = tactic -> tactic

a tactic_combinator is a function between tactics. It has many uses like sequencing tactics (then_one, then_all), handling specific effects (with_no_trace, with_fuel_limit), or managing search over a tactics choices (with_dfs, with_best_first.

type cost =

| Safe of int

| Unsafe of int

type choice_kind =

| CTerm of goal * Kernel.term

| CTheorem of goal * Kernel.thm

| CTactic of goal * cost * tactic

| CUnknown of goal

type search_metadata =

| MSubgoal of goal

| MChoice of choice_kind

| MResume

search_metadata is used by with_best_first to sort a priority queue, deciding which path of a proof space to explore next

type step_result =

| Cont of (choice_kind * (unit -> step_result)) list

| Need of goal * Kernel.thm -> step_result

| Done of Kernel.thm

| Dead

in search tactic_combinators, step_result is used to represent possible continuations of a search

module Priority : sig ... end

module PriorityQueue : sig ... end

module type Frontier = sig ... end

type _ rankable =

| Term : Kernel.term list -> Kernel.term rankable

| Goal : goal list -> goal rankable

| Tactic : tactic list -> tactic rankable

| Unknown : 'a list -> 'a rankable

the rankable GADT is used to allow both agnostic treatment of the Rank effect as well as deeper introspection into the underlying data when needed

type _ choosable =

| Term : Kernel.term list -> Kernel.term choosable

| Theorem : Kernel.thm list -> Kernel.thm choosable

| Tactic : tactic list -> tactic choosable

| Unknown : 'a list -> 'a choosable

the choosable GADT is used to allow both agnostic treatment of the Choose effect as well as deeper introspection into the underlying data when needed

exception Out_of_fuel

Out_of_fuel is performed by the with_fuel_limit tactic_combinator to indicate that a tactic has gone over its limit

type Stdlib.Effect.t +=

| Subgoal : goal -> Kernel.thm Stdlib.Effect.t

| Choose : 'a choosable -> 'a Stdlib.Effect.t

| Rank : 'a0 rankable -> 'a0 list Stdlib.Effect.t

| Fail : 'a1 Stdlib.Effect.t

| Trace : (level * string) -> unit Stdlib.Effect.t

| Burn : (string * cost) -> unit Stdlib.Effect.t

| Rules : Kernel.thm list Stdlib.Effect.t

val as_ranked_list : 'a. 'a rankable -> 'a list

as_ranked_list extracts the underlying type from the rankable GADT

val as_chosen_list : 'a. 'a choosable -> 'a list

as_chosen_list extracts the underlying type from the choosable GADT

step performs one expansion of the proof tree and aggregates the results along with their continuations

val cost_of_tactic : tactic -> goal -> string * cost

val step : tactic -> goal -> step_result

val fail : unit -> 'a

fail performs the Fail effect. This is used to signal when a tactic doesn't apply or doesn't make progress

val burn : string -> cost -> unit

burn performs the Burn effect. This is used to signal the 'cost' of a tactic relative to other tactics

val trace_dbg : string -> unit

trace_dbg emits a debug-level trace message

Effects: Trace

val trace_info : string -> unit

trace_info emits an info-level trace message

Effects: Trace

val trace_error : string -> unit

trace_error emits an error-level trace message

Effects: Trace

val trace_proof : string -> unit

trace_proof emits a proof-level trace message, used by tactics to record their name in the proof path

Effects: Trace

val choose_terms : Kernel.term list -> Kernel.term

choose_terms requests a choice among a list of terms

Effects: Choose

val choose_theorems : Kernel.thm list -> Kernel.thm

choose_theorems requests a choice among a list of theorems

Effects: Choose

val choose_tactics : tactic list -> tactic

choose_tactics requests a choice among a list of tactics

Effects: Choose

val choose_unknowns : 'a list -> 'a

choose_unknowns requests a choice among a list of unknown type

Effects: Choose

val rank_terms : Kernel.term list -> Kernel.term list

rank_terms requests a ranking/sorting of terms by some heuristic

Effects: Rank

val return_thm : ?from:string -> ('a, [< `BadSubstitutionList | `CantApplyNonFunctionType of 'b | `CantCreateVariantForNonVariable of 'c | `CantDestructEquality | `Clash of 'd | `ConstantTermAlreadyDeclared of 'e | `ConstructorsAlreadyExist | `DefinitionError of 'f | `Eq_MP of 'g | `InvariantViolation of string | `LamRuleCantApply | `MakeAppTypesDontAgree of Kernel.hol_type * Kernel.hol_type | `MakeLamNotAVariable of 'h | `NameMappingError of 'i | `NewAxiomNotAProp | `NewBasicDefinition of 'j | `NewBasicDefinitionAlreadyDefined of 'k | `NoBaseCase | `NoRewriteMatch | `NotAConj | `NotAConst | `NotAConstantName of 'l | `NotADisj | `NotAForall | `NotALam | `NotANegation | `NotAProposition | `NotAVar | `NotAnApp | `NotAnApplication of 'm | `NotAnExists | `NotAnImp | `NotBothEquations | `NotFreshConstructor | `NotPositive | `NotTrivialBetaRedex | `OperationDoesntMatch of 'n | `RuleTrans | `Todo | `TypeAlreadyDeclared of 'o | `TypeAlreadyExists | `TypeConstructorNotAVariable of 'p | `TypeDefinitionError of 'q | `TypeEquivalenceNotImplemented | `TypeNotDeclared of 'r | `TypeVariableNotAConstructor of 's | `TypesDontAgree | `UnexpectedLambdaForm | `WrongNumberOfTypeArgs of 't ]) Stdlib.result -> 'a

return_thm is used by tactics to handle failure and trace information about which tactic was run

val left_tac : tactic

left_tac takes goals like P \/ Q and creates the subgoal P. It fails if the goals conclusion is not a conjunction. This tactic is not safe, as it is not true that P is always provable when P \/ Q is

Effects

Fail

Subgoal

Burn

Trace

val right_tac : tactic

right_tac takes goals like P \/ Q and creates the subgoal Q. It fails if the goals conclusion is not a conjunction. This tactic is not safe, as it is not true that Q is always provable when P \/ Q is

Effects

Subgoal

Burn

Fail

Trace

val or_tac : tactic

or_tac chooses between left_tac and right_tac, ensuring both sides are attempted if used under a search combinator

Effects: Choose, Fail, Burn, Trace

val apply_asm_tac : tactic

apply_asm_tac finds assumptions of the form P -> Q for the goal Q and creates a subgoal P

Effects: Rank, Choose, Burn, Fail, Subgoal

val apply_thm_tac : tactic

apply_thm_tac applies a chosen theorem from Rules by stripping foralls and matching the conclusion against the goal

Effects: Rules, Choose, Burn, Trace, Subgoal, Fail

val apply_thm_asm_tac : tactic

apply_thm_asm_tac applies a chosen theorem from Rules to a chosen assumption. If theorem is P ==> Q and assumption is P, replaces the assumption with Q and creates a subgoal with the updated assumptions

Effects: Rules, Choose, Burn, Trace, Subgoal, Fail

val apply_neg_asm_tac : tactic

apply_neg_asm_tac proves F by finding a negation ~P in assumptions and creating a subgoal to prove P. Fails if the goal is not F or no suitable negation exists

Effects: Choose, Burn, Trace, Subgoal, Fail

val assume_tac : tactic

assume_tac proves any goal by assuming it. This creates a theorem with the goal as a hypothesis

Effects: Burn, Trace

val sym_tac : tactic

sym_tac transforms a goal l = r into r = l

Effects: Burn, Trace, Subgoal, Fail

val rewrite_tac : tactic

rewrite_tac rewrites the goal using a chosen theorem from Rules. Performs subterm matching and fails if no progress is made

Effects: Rules, Choose, Burn, Trace, Subgoal, Fail

val rewrite_asm_tac : tactic

rewrite_asm_tac rewrites a chosen assumption using a theorem from Rules. Fails if no progress is made

Effects: Rules, Choose, Burn, Trace, Subgoal, Fail

val beta_tac : tactic

beta_tac performs deep beta reduction on the goal and creates a subgoal with the reduced term

Effects: Burn, Trace, Subgoal, Fail

val beta_asm_tac : tactic

beta_asm_tac performs deep beta reduction on a chosen assumption. Fails if no progress is made

Effects: Choose, Burn, Trace, Subgoal, Fail

val mp_asm_tac : tactic

mp_asm_tac finds an implication P ==> Q in assumptions where P is also an assumption, and adds Q to the assumptions. Fails if no such implication exists or Q is already present

Effects: Choose, Burn, Trace, Subgoal, Fail

val intro_tac : tactic

intro_tac transforms a goal P ==> Q into a subgoal Q with P added to the assumptions. Fails if goal is not an implication

Effects: Burn, Trace, Subgoal, Fail

val refl_tac : tactic

refl_tac proves goals of the form t = t by reflexivity. Fails if the goal is not an equality or the sides are not identical

Effects: Burn, Trace, Fail

val assumption_tac : tactic

assumption_tac proves the goal if it matches one of the assumptions. Fails if no matching assumption is found

Effects: Choose, Burn, Trace, Fail

val conj_tac : tactic

conj_tac transforms a goal P /\ Q into two subgoals P and Q. Fails if the goal is not a conjunction

Effects: Burn, Trace, Subgoal, Fail

val elim_disj_asm_tac : tactic

elim_disj_asm_tac eliminates a disjunction P \/ Q in the assumptions by case splitting, creating two subgoals: one with P and one with Q

Effects: Choose, Burn, Trace, Subgoal, Fail

val elim_conj_asm_tac : tactic

elim_conj_asm_tac eliminates a conjunction P /\ Q in the assumptions by replacing it with both P and Q as separate assumptions

Effects: Choose, Burn, Trace, Subgoal, Fail

val neg_elim_tac : tactic

neg_elim_tac proves any goal when both P and ~P are in assumptions, deriving a contradiction. Fails if no such pair exists

Effects: Choose, Burn, Trace, Fail

val neg_intro_tac : tactic

neg_intro_tac transforms a goal ~P into a subgoal F with P added to the assumptions. Fails if goal is not a negation or P is already an assumption

Effects: Burn, Trace, Subgoal, Fail

val ccontr_tac : tactic

ccontr_tac proves P by classical contradiction: assumes ~P and derives F. This is a classical (non-intuitionistic) tactic

Effects: Burn, Trace, Subgoal, Fail

val false_elim_tac : tactic

false_elim_tac proves any goal when F (false) is in the assumptions. Fails if F is not present

Effects: Burn, Trace, Fail

val gen_tac : tactic

gen_tac transforms a goal forall x. P into a subgoal P. Fails if the goal is not a universal quantification

Effects: Burn, Trace, Subgoal, Fail

val induct_tac : tactic

induct_tac applies structural induction on the quantified variable of a goal forall x. P. Creates subgoals for each constructor case of the inductive type. Fails if the goal is not universally quantified or the type is not inductive

Effects: Burn, Trace, Subgoal, Fail

val prove : ?name:string -> goal -> tactic -> proof_state

prove is the main effect handler that runs a tactic on a goal. It provides default interpretations for all effects: printing traces, taking first choices, ignoring fuel costs, etc. Returns Complete thm on success or Incomplete goal on failure

val then_one : tactic -> tactic_combinator

then_one sequences two tactics: applies tac1 then applies tac to only the first subgoal. Remaining subgoals bubble up. Infix: >>

val (>>) : tactic -> tactic_combinator

val then_all : tactic -> tactic_combinator

then_all sequences two tactics: applies tac1 then applies tac to all subgoals. Infix: >>>

val (>>>) : tactic -> tactic_combinator

val then_each : tactic list -> tactic_combinator

then_each applies a list of tactics to subgoals in order. Fails if there are more subgoals than tactics provided. Infix: >>=

val (>>=) : tactic -> tactic list -> tactic

val with_first_success : tactic_combinator

with_first_success handles Choose by trying each choice in order until one succeeds. Only handles choices at one level; for recursive search use with_dfs or with_best_first

val with_term : Kernel.term -> tactic_combinator

with_term forces a specific term to be chosen when a Choose (Term _) effect is performed. Fails if the term is not among the choices

val try_ : tactic_combinator

try_ converts failure into a subgoal request, a tactics sequence to be used in situations where one or more intermediate tactics could fail

val pick_tac : tactic list -> tactic

pick_tac creates a tactic that chooses among the given tactics. Used with search combinators to explore different proof strategies

Effects: Choose

val solve : tactic_combinator

solve requires a tactic to completely solve the goal without leaving subgoals. Fails if any subgoals remain

val run_thunk_with_path : string list Stdlib.ref -> (unit -> 'a) -> 'a

run_thunk_with_path executes a thunk while capturing Trace (Proof, _) effects into the provided path reference. Used by search combinators to track the winning proof sequence

val emit_proof_path : string list -> unit

emit_proof_path formats a proof path as a tactic script and emits it as a Trace (Search, _) effect

Effects: Trace

val stats_of_list : (search_metadata * 'a * 'b * 'c) list -> string

module StackFrontier : Frontier

val make_search : (module Frontier) -> tactic_combinator

val with_dfs : tactic_combinator

with_dfs performs depth-first search over choices and subgoals. Explores the proof space using a stack, backtracking on failure. Emits the winning proof path on success

Effects: Trace (Search) on success, Fail if no proof found

module PQueueFrontier : Frontier

val with_best_first : tactic_combinator

with_best_first performs best-first search over choices and subgoals. Uses a priority queue ordered by search_metadata to explore promising paths first (resumes before choices, choices before subgoals). Emits the winning proof path on success

Effects: Trace (Search) on success, Fail if no proof found

module QueueFrontier : Frontier

val with_bfs : tactic_combinator

with_bfs performs breadth-first search over choices and subgoals. Uses a queue to explore paths level by level. Emits the winning proof path on success

Effects: Trace (Search) on success, Fail if no proof found

val with_dfs'' : tactic_combinator

with_dfs'' is an alternative DFS implementation using an explicit stack for choice points. Does not track proof paths

val with_dfs' : tactic_combinator

with_dfs' is a recursive DFS implementation that uses the call stack for backtracking. Simpler but may overflow on deep searches. Does not track proof paths

val with_repeat : tactic_combinator

with_repeat repeatedly applies a tactic until it fails or makes no progress. On failure after progress, emits a subgoal for the current state

val itauto_tac : tactic

itauto_tac is a complete automation tactic for intuitionistic propositional logic. Chooses among various introduction and elimination tactics. Use with a search combinator like with_dfs or with_best_first

Effects: Choose

val ctauto_tac : tactic

ctauto_tac is a complete automation tactic for classical propositional logic. Chooses among various introduction and elimination tactics. Use with a search combinator like with_dfs or with_best_first

Effects: Choose

val ctauto_dfs_tac : tactic

ctauto_dfs_tac is ctauto_tac wrapped with with_dfs for automatic depth-first proof search

val with_interactive_choice : tactic_combinator

with_interactive_choice handles Choose effects by prompting the user to select from the available options via stdin

val with_nth_choice : int -> tactic_combinator

with_nth_choice n always selects the nth option from any Choose effect. Fails if n is out of bounds

val with_term_size_ranking : tactic_combinator

with_term_size_ranking handles Rank (Term _) effects by sorting terms from smallest to largest based on AST size

val cost_value : cost -> int

val with_added_fuel : int -> tactic_combinator

val with_fuel_limit : int Stdlib.ref -> tactic_combinator

with_fuel_limit tracks fuel consumption and raises Out_of_fuel when the limit is exceeded. The limit is a mutable reference that decreases with each Burn effect

val with_fuel_counter : int Stdlib.ref -> tactic_combinator

with_fuel_counter tracks total fuel consumed by incrementing a mutable reference for each Burn effect

val with_no_trace : ?show_proof:bool -> tactic_combinator

with_no_trace suppresses trace messages. By default suppresses all except Search. Set show_proof:true to also show Proof traces

val with_assumptions : tactic_combinator

with_assumptions provides the goal's assumptions as theorems when a Rules effect is performed

val with_rules : Kernel.thm list -> tactic_combinator

with_rules provides a fixed list of theorems when a Rules effect is performed

val with_rule : Kernel.thm -> tactic_combinator

with_rule provides a single theorem when a Rules effect is performed

val with_proven : string list -> tactic_combinator

with_proven looks up previously proven theorems by name and provides them when a Rules effect is performed. Fails if any name is not found

val with_rules_and_assumptions : Kernel.thm list -> tactic_combinator

with_rules_and_assumptions provides both the given rules and the goal's assumptions as theorems when a Rules effect is performed

val intros_tac : tactic

intros_tac repeatedly applies intro_tac and gen_tac until neither makes progress. Useful for introducing all hypotheses at once

val simp_tac : ?with_asms:bool -> tactic

simp_tac simplifies the goal using rewrite rules from definitions and registered simp lemmas. Set with_asms:false to exclude assumptions

Effects: Rules, Choose, Burn, Trace, Subgoal, Fail

val auto_tac : tactic

auto_tac is an automation tactic combining simplification with basic logical tactics. Use with a search combinator for full automation

Effects: Choose

val auto_dfs_tac : tactic

auto_dfs_tac is auto_tac wrapped with with_dfs for automatic depth-first proof search

val simp_asm_tac : ?with_asms:bool -> ?add:Kernel.thm list -> tactic

simp_asm_tac simplifies assumptions using rewrite rules from definitions. Use add to provide additional rules. Set with_asms:false to exclude other assumptions as rewrite rules

Effects: Rules, Choose, Burn, Trace, Subgoal, Fail

val run_proof : ?notrace:bool -> ?name:string -> goal -> tactic -> unit